Add a stylesheet to harvest training data
Run it something like:

  cd xslt

  ./harvest_trainingdata.sh trainingdata.tab '../../eebotcp'

It requires Java and Saxon.  The file produced is in the same
format used by createlexicon except it has an extra column at the
end contaning the year. For now we're assuming the year can be
found in places that can be depended on in the EarlyPrint Library,
but this could be adapted pretty easily to any location that can
be relied on to have the year of the text's production.

The expecation is that the year will be used as a filter to
produce training data sets appropriate to a particular date range.
The filtering process should remove the year column but otherwise
pass through the remainder of the relevant records. A line that has
empty columns except for the year should produce an empty line in
the output as this is used for sentence endinds in subsequent
processing steps.

There are a number of ways this could be done, but one easy one is:

  perl -F'\t' -anE \
   'print ((length($F[0]) == 0 ? "" : (join qq/\t/, @F[0,1,2,3])) . qq/\n/) if $F[4] < 1580;' \
   < trainingdata_by_year.tab \
   > trainingdata_before_1580.tab

where that example creates a set of training data for lexical
items occurring before 1580.  That resulting file would then be
used to create period-specific spelling and suffix lexica like
so:

  cd ..

  ./createlexicon xslt/trainingdata_before_1580.tab \
  data/wordlexiconeme_before_1580.lex \
  data/suffixlexiconeme_before_1580.lex

and tag transition possibility matrices like so:

  ./ngramtaggertrainer xslt/trainingdata_before_1580.tab \
  data/wordlexiconeme_before_1580.lex \
  data/emetransmat_before_1580.mat