Omit "zz" POS tags from the training data
These mean "unknown" but the training data is supposed to provide
hints based on things we know. The presence of these entries in
the lexica generated from the training data means that sometimes
zz gets picked as the likeliest part of speech, but that's turning
a negative assertion into a positive one.

So omit this tag from the training data to avoid the propagation
of disknowledge. MorphAdorner will still choose zz sometimes when
all of its heuristics fail to find anything else, but that's an
appropriate use of the tag.