Update NCF and ECCO lexica to NUPOS v3
Many, many part-of-speech tags have evolved over the years, mostly
in the direction of simplification, but leaving quite a few orphans
in their wake.  This represents a best effort at translating all
the variants into their current form.  There are a handful that are
left untranslated (but preserved); most of these were never real
tags to begin with.